Corrects implementation of removeWatch (issue #11)
As @jafingerhut pointed out, there are two apparently incorrect uses
of swap!, one in atoms.clj and one in refs.clj. In both cases, the
error is in the removeWatch implementation (shown to be broken by the
new tests in my previous commit). This commit fixes resolves #11 and
resolves #3. I can't tell for sure yet, but I suspect that it might
help with #10.